Extract alfa-table package
