--- packages/alfa-dom/src/tables/groups.ts
@@ -1,5 +1,0 @@
-// groups of slots in a table
-export * from "./colgroup";
-export * from "./cell";
-export * from "./row";
-export * from "./rowgroup";
