+++ packages/alfa-table/package.json
@@ -1,6 +1,6 @@
 {
   "private": true,
-  "name": "@siteimprove/alfa-aria",
+  "name": "@siteimprove/alfa-table",
   "author": "Siteimprove A/S",
   "homepage": "https://siteimprove.com",
   "version": "0.0.0",
@@ -18,26 +18,19 @@
   "main": "src/index.js",
   "types": "src/index.d.ts",
   "files": [
+    "*.js",
+    "*.d.ts",
     "src/**/*.js",
     "src/**/*.d.ts"
   ],
   "dependencies": {
-    "@siteimprove/alfa-branched": "^0.0.0",
-    "@siteimprove/alfa-cache": "^0.0.0",
-    "@siteimprove/alfa-compatibility": "^0.0.0",
-    "@siteimprove/alfa-device": "^0.0.0",
+    "@siteimprove/alfa-comparable": "^0.0.0",
     "@siteimprove/alfa-dom": "^0.0.0",
     "@siteimprove/alfa-equatable": "^0.0.0",
     "@siteimprove/alfa-iterable": "^0.0.0",
     "@siteimprove/alfa-json": "^0.0.0",
-    "@siteimprove/alfa-lazy": "^0.0.0",
-    "@siteimprove/alfa-map": "^0.0.0",
-    "@siteimprove/alfa-mapper": "^0.0.0",
     "@siteimprove/alfa-option": "^0.0.0",
-    "@siteimprove/alfa-predicate": "^0.0.0",
-    "@siteimprove/alfa-sequence": "^0.0.0",
-    "@siteimprove/alfa-style": "^0.0.0",
-    "@siteimprove/alfa-table": "^0.0.0"
+    "@siteimprove/alfa-predicate": "^0.0.0"
   },
   "devDependencies": {
     "@siteimprove/alfa-test": "^0.0.0"
