+++ packages/alfa-table/src/index.ts
@@ -1,0 +1,3 @@
+export * from "./header";
+export * from "./table";
+
