+++ packages/alfa-table/src/rowgroup.ts
@@ -1,12 +1,13 @@
 import { Comparable } from "@siteimprove/alfa-comparable";
+import { Element, isElementByName } from "@siteimprove/alfa-dom";
 import { Equatable } from "@siteimprove/alfa-equatable";
 import { Serializable } from "@siteimprove/alfa-json";
 import { Err, Ok, Result } from "@siteimprove/alfa-result";
 
 import * as json from "@siteimprove/alfa-json";
 
-import { Element, isElementByName } from "..";
-import { Row, Cell } from "./groups";
+import { Cell } from "./cell";
+import { Row } from "./row";
 
 const { compare, Comparison } = Comparable;
